feat: added TensorFlow train script
Now you can run 'python train.py' to load pre-trained model & train
on it some more.  There's an option to re-initialize the weights and
train from scratch, but the argument parser is not implemented yet
and will come later.
Updated README to reflect the changes.
Sloppy on documentation, but will update soon.